requirements.txt
aiosqlite
discord.py
openai
Requests
python-dotenv
PyNaCl